requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-stubs==1.1.0
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2
MarkupSafe==1.1.1
maxminddb==1.5.2
openapi-codec==1.3.2
psycopg2==2.8.4
pytz==2019.3
PyYAML==5.3
requests==2.22.0
simplejson==3.17.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
